site/server/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Auth server"
authors = ["Fabián <fabianveal@gmail.com>"]
readme = "README.md"
packages = [
    { include = "auth_server", from = "." },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
django-oauth-toolkit = "^2.3.0"
django-cors-headers = "^4.2.0"
grpcio-tools = "^1.59.0"


[tool.poetry.group.format.dependencies]
black = "^23.9.1"
isort = "^5.12.0"


[tool.poetry.group.lint.dependencies]
ruff = "^0.0.291"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
manage = "manage:main"
